risch: add link to the Poor Man's Integrators and original tests
It seems our risch_norman is very close to the Poor Man's Integrator (PMINT) as
written by Emanuel Bronstein. Thus I decided to add link to PMINT and add
original tests.

Unfortunately the tests do not pass now, but I think it is nontheless right to
include them to know where to work in the future.