Associated Legendre polynomials.
We define them through recurrence relation. Maybe it is not the best way to go,
but at least it does fit to how other polynomials are defined.

Let's have it consistent?

----

Unfortunately, without simplify I failed to make the result of assoc_legendre
pretty as commonly shown in the books and wikipedia, and with simplify it is
pretty, but the performance is bad.

So, instead of using reccurence formula, I used the direct one. The performance
still benefits from memoization.