REPR0 decorator for test
rationale:
----------

In tests, we need to check the behaviour of real __repr__ oftenly, but we set
Basic.set_repr_level(1) on sympy import time, so __repr__ goes to __str__,
which leads to confusion.

To avoid this in tests, I propose we use REPR0 decorator for functions needing
repr & str, and just use repr(...) and str(...) there.