TeXmacs interface (#469)
Currently it only does not support plotting directly to a
TeXmacs document (Plot.saveimage() returns empty image for me).

To use this copy the contents of data/TeXmacs to the tar-ball to
~/.TeXmacs/plugins

TeXmacs should automatically add SymPy to Insert/Session
menu. If not, select Other and type sympy.

tm_sympy imports sympy * by default, creates Greek symbols and a few Latin.
gamma(x) and zeta(x) are renamed to Gamma(x) and Zeta(x) respectively to avoid
symbol redefinition.

Added data/* to manifest so that tm_sympy will be inclued in sdist.