[1/4] Namespace for SymPy classes
As was shown before calling Basic.<smth> to get that "something" class is
expensive. It would be better to explicitly import everything we need, but the
so-called cyclic-import problem arises, so we have to do some gymnastics to avoid it.

The first idea is to cache results in Basic.__dict__, so that __getattr__ is
called only once and next accesses are faster, but  it wouldn't be good if say
x.Rational will be meaningful, and we explicitly stated in our testcases this
is not allowed.


So, I think the time has come to move global class namespace from Basic to a
place specially prepared for it -- if we have `S` for singletons, lets have `C`
for classes?

Accessing classes is like before:

  C.<class_name>

e.g.

  C.Rational
  C.Add


Only C can cache the results, so the time for attribte access is approximately
the same as for dict item access:

  %timeit Basic.Rational    11.2 µs
  %timeit C.Rational         624 ns

   ratio:                   ~ 20x


This patch is just for ClassesRegistry -- C's engine. Patches that adjust SymPy
all over the place will follow.