[1/7] _sympifyit -- fast decorator to _sympify(arg)
This is `_sympify` -- shorter version of `sympify` as discussed in (#677).
`_sympify` accepts only Basic, int/long/float/Decimal/complex and instances with
._sympy_ . Everithing else is reported as an error, so

  x + '1'

raises.

Also, included is smart decorator `_sympifyit`. It's task is to get all the
tedious work to inject _sympifyit into the process and not degrate performance.

We'll switch ArithMeths and RelMeths to it in the next commits.

Also, in the end of this patch series, there is a big surprise that (again)
speeds up `_sympifyit`. Stay tuned.