[5/7] _sympifyit -- fast decorator to _sympify(arg)
Switch numbers.py to `_sympifyit` & `_sympify`.

I have to say that numbers need a lot of work and a lot can be potentially
optimezed there. This change only uses `_sympify` instead of `sympify` and
wraps Rational __add__ and __mul__ with `_sympifyit`

Timings
-------

r = Rational(3,4)
q = Rational(5,4)

        %timeit  %timeit
         r + 1    r + q

old:     138 µs   112 µs
new:     131 µs   112 µs

As we can see Rationals are *slow* -- this should be the task for future
improvements.