[2/5] cache cleanup, hadle only immutable objects
Now we've eliminated all usage of cache_it, let's kill to what it pointed --
cache_it_fast.

Plus kill mycopy (deep-coping)

Correctnes of the whole thing was verified with

    SYMPY_USE_CACHE=debug py.test sympy/core/

and

    SYMPY_USE_CACHE=debug py.test sympy/


but in the second check cache_it_debug was modified to only check that the
result is immutable. Otherwise we run out of 1G of memory.