Fix tests to be ordering (in the Add class) neutral.
This patch makes all tests not depending on the way Add stores it's terms
internally and also how str() then outputs the result. It probably doesn't
cover 100% of cases, but it's a notable improvement over the past.

This patch also divides some longer test functions into more smaller ones,
because some of them break easily when we change the terms ordering in Add,
so that it's easy to XFAIL them later if we have to. Also it's good for
debugging purposes. Clearly those tests are special in some way, so they should
imho be in their own test functions.