[1/2] @cacheit Wild
We use Wilds a lot internally, and each new Wild creates another new object.

Wilds are usually then put into match or subs, and since match and subs are
cached, we have performance penalty:

caching only slows things down in this case  -- we just have overheads (one
more call, always store the result in the cache) but never use the cache, so
the cache just blows, caching slows things down and we get nothing back.


So let's just cache Wilds (and btw disallow Wild())

If we are running under SYMPY_USE_CACHE=no we are not loosing anything, but if
we are running with cache turned on, we gain significant speedup:

        %time integrate(x*sin(x), x)    py.test
          1st run / other runs          sympy/

old:      2.26s     2.18s               454s
new:      2.01s     1.89s               404s