[1/5] __eq__/__ne__ returns True/False directly  (#135)
Previously we used to create Relational instances in this methods, but this
comes with performance penalty. And since dict uses __eq__ and __ne__ for hash
collision resolution this is very important.

So let's return True/False directly. We'll adapt the rest of SymPy in the
following patches. For you to stay tuned, here are timings we'll achieve in the
end.

Timings
-------
        bool(x==y) integrate(x*sin(x), x) fem_test  py.test      py.test
                                          fem_test  sympy/core/  sympy/
old:    ~200us         1.90s                5.33s     12.4s       370s
new:     ~18us         1.33s                4.65s     11.6s       295s

speedup: >1000%        42%                  14%       6%          25%