Use limit() in Basic.limit(), instead of limit_series().
Note: the Basic.inflimit() will be removed in the next patch -- that's because
the test_limits_series.py test still uses it (still 5 tests fail in
test_limits_series.py, but I left it failing, as the file will be removed in
the next patch anyway).

Also the "wsym" temporary variable in the gruntz algorithm was set as Real,
otherwise a few series tests failed. With this, no test fail (except 5 in
test_limits_series.py).

Speed
-----

Before: 31s
Now: 42s

But that's imho due to using limits in .oseries() too much (oseries is still
called from nseries for some special cases). Consider:

$ py.test sympy/series/tests/test_nseries.py
before: 7.43s
now:   11.57s

And in partucular:

$ py.test sympy/series/tests/test_nseries.py -k expbug4
before: 0.50
now:    1.34

So .nseries() of exp ang log functions need to be improved.