Use mpmath instead of the decimal module.
Nearly all uses the decimal module have been replaced by mpmath. In
particular, the Real class uses an mpf to hold its value, and functions
are now evaluated using mpmath.

There is only one change in behavior: Real(2.0) now becomes Real(2.0)
instead of Integer(2). This only broke a single test, and this test
was designed specifically to look for that case. Another test was broken
in spherical harmonics, but a more robust fix in that case was to sympify
the input (now fixed).

The exception is Real(0.0), which still becomes Integer(0), since the
code that checks for symbolic cancellation seemingly assumes a single
0 instance. There are some more subtle issues involved here, and the
behavior of Real() will have to be revised in the future, but the
present behavior was the easiest solution for now and it doesn't
break anything.

Default precision is now 15 digits (same as mpmath) instead of 28
(as with decimal). This would be trivial to change back. The epsilon
in a polynomial root test was changed to account for the reduced
precision.

decimal_math.py has been deleted since it is redundant.