doc: spt -- kill duplicate leftover after a4ab1c5072c4 and 1e371247bb0b
Ondrej, at sympy-patches I asked you whether it makes sense to move `Quick
Start` section to appendix:

  http://groups.google.com/group/sympy-patches/browse_thread/thread/d5a132b22db0675c/9e5ef081fd2de438#9e5ef081fd2de438

And it seems you agree with this idea:

  http://groups.google.com/group/sympy-patches/msg/769a69514a904897

But instead of putting the text into one patch (which do one thing and do it
well) you decided to do it in two steps:

  a4ab1c5072c4 and 1e371247bb0b

But the problem is that second commit 'forgets' to move some text left after
the first one.

I think this is general -- when we prepare/review patches it is better to first
get them into shape, and only then merge.

Only if you are a big-subsystem maintanier, like say linux-scsi it makes sense
to keep patches intact once merged, so others could base their work on yours,

But in our case, where we have ~ 10 patches pending this is harmful, because
we'd better prepare that patches in shape first to keep the quality, and for
others to use your work, publishing repository with your MQ does work.

Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>