Fixes sin(-x-1) on a different python platform.
On amd64 there is a different ordering of terms, so this patch is needed there.
What's wrong without the patch -- those two tests fail:

   assert sin(-x-1) == -sin(x+1)
   assert cos(-x-1) == cos(x+1)

and that is because on i386:

In [1]: (-x-1).args
Out[1]: (-1, -x)

and amd64:

In [1]: (-x-1).args
Out[1]: (-x, -1)

So when we look at the code in sympy/functions/elementary/trigonometric.py:

                   if arg.args[0].is_Mul:
                       if arg.args[0].args[0].is_negative:
                           # e.g. arg = -x - y
                           if (-arg).args[0].is_Mul:
                               if (-arg).args[0].args[0].is_negative:
                                   # This is to prevent infinite recursion in
                                   # the case cos(-x+y), for which
                                   # -arg = -y + x. See also #838 for the
                                   # root of the problem here.
                                   return
                           # convert cos(-x-y) to cos(x+y)
                           return cls(-arg)
                   if arg.args[0].is_negative:
                       return cls(-arg)

We can see, that arg.args[0] returns -x on i386, and so the "return cls(-arg)" statement gets executed, while on amd64 it return -1, so nothing gets executed without the patch. Adding the last two lines fixes that.

A general fix would be to always order the arguments of Add/Mul in the same way
independently of the platform, but that requires some deep refactoring and also
I think it's not bad that SymPy is able to work with any ordering.

Also a test for infinite recursion had to be added.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>