Raise SymbolsError in Poly.__init__ and allow Poly(Poly(x, x))
To distinguish between types of erros in polys module, new error is
introduced: SymbolsError. This is needed to clarify the interface.

To make Poly usage simpler (e.g. this will be used in Poly.cancel)
one don't have to specify symbols if the main argument is itself
a Poly. In this case it is a quick return (for speed).