Use lists as results rather than tuples in rootfinding.py
Using tuples as results in functions like roots() (with multiple
flag set), factors or sturm, was generally a bad idea (this was
not a good place to use tuples and for consistency with other
modules, e.g. solvers, this had to be changed). The same will
be done with functions in algorithms.py