New syntax in Poly.cancel, allow (numer, denom) tuples
In situations where we have already the numerator and denominator
separated, there is no more need to combine those expressions in
to a single fraction (improved speed).

Previously:

 Poly.cancel(numer/denom, x, y)

Currently:

 Poly.cancel((numer, denom), x, y)

Note that numer and denom can be Polys. In this case there is no need to
specify symbols. Otherwise symbols MUST be given explicitely, as atoms()
is not used.

Semantics of Poly.cancel(f, x, y) remains the same.