Fixes pretty printing/isympy on windows (#933).
It turns out that the windows curses in my python distribution (ActiveState)
doesn't implement setupterm or tigetnum.  So, the terminal_string method in
stringpict.py fails when trying to get the terminal width.  I used a recipe
from the ASPN cookbook to fix this:

http://aspn.activestate.com/ASPN/Cookbook/Python/Recipe/440694

The following example works now:

In [1]: (1/cos(x)).series(x, 0, 10)
Out[1]:
     2      4       6        8
    x    5*x    61*x    277*x
1 + -- + ---- + ----- + ------ + O(x**10)
    2     24     720     8064