Fixed bug in Poly.evaluate
Now evaluate() will return non Poly result when the number
of evaluation symbols is equal to the number of symbols in
the input polynomial, even if the result is a constant
polynomial (e.g. Poly(0, x)).

Also an invalid (0, (x, y, z)) term won't appear in the
output. This bugfix is a result of work on poly solver.