Improve simplification of radicals in Poly.cancel
This patch is a quick hack which improves simplification of trivial
radicals in Poly.cancel and it makes possible of computing e.g:

In [1]: solve_poly_system([x**2+y+z-1, x+y**2+z-1, x+y+z**2-1], x, y, z)
Out[1]:
⎡⎛       ⎽⎽⎽         ⎽⎽⎽         ⎽⎽⎽⎞                        ⎛       ⎽⎽⎽         ⎽⎽⎽         ⎽⎽⎽⎞           ⎤
⎣⎝-1 - ╲╱ 2 , -1 - ╲╱ 2 , -1 - ╲╱ 2 ⎠, (0, 0, 1), (0, 1, 0), ⎝-1 + ╲╱ 2 , -1 + ╲╱ 2 , -1 + ╲╱ 2 ⎠, (1, 0, 0)⎦

However this does not solve a real problem, which will require quite
some effort to fix. When doing polynomial computations we should do
out best to keep coefficients as small as possible. Even in case of
integer polynomials this is not a trivial task and modular approach
is used in this case. If we have more complicated coefficients, e.g.
algebraic numbers, rational functions (possibly multivariate), then
it requires even more effort to keep them small.

We could just apply all available transformations (simplifications)
but then we would degenerate speed and even a simple division would
be a nightmare process.

So to solve this in general we should leave the idea that coefficients
can be whatever. Each polynomial must have explicitely and precisely
declared type of its coefficients, e.g:

 [1] integers
 [2] rationals
 [3] algebraic numbers
 [4] polynomials (UV, MV)
 [5] rational functions (UV, MV)
 [6] general expressions
  .
  .
  .
 [oo] whatever user likes

Then each polynomial over a different coefficient domain would be a
different type, with a set of special functions implemented. One of
them is SomePoly.reductum (curretly Poly.cancel). Example of other
function of this kind is divide_if_can(), which would check its
arguments and return a result of exact quotient or None (if fact
this should be implemented in the core, but for now it is sufficient
to have it in polynomials.

If user would like to have its own type of polynomial then he will
create a new poynomial type and implement new reduction method.

This way we will have full control on what happens with coefficients
when being processed (mostly when division occurs, but this may be
hidden in other arthmetic operations (consider rational functions)),
and we will have speed, because no time will be lost on useless
simplifications.

This will also close issues like #326 once and for all.