BasicMeta: don't import typeinfo (e.g. is_Integer) into class default_assumptions
Because this was slowing down things significantly:


        %timeit x+y     fem_test.py     fem_test.py     py.test sympy/core/
        (cache: off)    (cache: off)    (cache: on)     (cache: on)

old:    436 µs          16.57 s         4.79 s          9.59 s
new:    297 µs          13.18 s         4.35 s          8.75 s

speedup: 47%             26%             10%             10%


In fact when we introduced is_XXX typeinfo (#748, f83106bc39a5) I've noticed
~10% performance regression:

    http://code.google.com/p/sympy/issues/detail?id=748#c15
    http://code.google.com/p/sympy/issues/detail?id=748#c18

but noone, including is_XXX original author, cared to study what the problem is.


I think it is bad practice to allow performance regressions without clear
understanding why such a change is needed and what benefits it brings.

When there is a regression we should try harder to understand what is going
wrong and what the problem is, and then the overal quality will improve.

In this particular case I was just enough lucky to detect this slowdown by an acident.


Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>