Removes the sys.path hack from sympy/__init__ (#925).
This is achieved by introducing import_thirdparty() function, that handles
imports of all thirdparty packages (currently only pyglet).

This is a step in the right direction, e.g now:

    >>> import sympy
    >>> import pyglet
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
    ImportError: No module named pyglet

As it should be. However, once you use plotting, the pyglet creeps back in:

    >>> import sympy
    >>> import pyglet
    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
    ImportError: No module named pyglet
    >>> import sympy.plotting.plot
    >>> import pyglet
    >>>

So this needs to be improved further. One idea is to use imputil + friends to
craft our own solution which does not populate sys.modules.