Fixes Order._cache bug (#948).
It took me 2 intensive hours to figure out what's wrong. Here is a script that
makes it fail:

    from sympy import Symbol, log, sin, O

    x = Symbol("x", real=True)
    w = Symbol("w", real=True)
    e = w**(1-log(x)/(log(2) + log(x)))
    f = e.series(w, 0, 1)
    e = sin(2*w)/w
    f = e.series(w, 0, 2)
    print "OK:", 2 + O(w**2)
    print "f :", f

On amd64 it prints (wrong):

    OK: 2 + O(w**2)
    f : O(w**2)

On i386 it prints (right):

    OK: 2 + O(w**2)
    f : 2 + O(w**2)


I will not describe the tedious way to figure out where the bug is, but here is
what goes wrong: the Order class uses cache for storing expression, see the
Order._get_cache_index()'s docstring. However, due to the poor and hackish
implementation of this method, when the cache contains the expression:

w**(1 - 1/(log(2) + log(x))*log(x))

it starts to add the rest of the terms wrong. E.g. it is something like this
[O(1/w), O(1), w**(1 - 1/(log(2) + log(x))*log(x)), O(w), O(w**2)], which is
correct, but then it adds w**3 in an utterly wrong position:

[O(1/w), O(1), O(w**3), w**(1 - 1/(log(2) + log(x))*log(x)), O(w), O(w**2)]

And so the rest of the algorithm returns wrong results. What is worse -- once
this cache is wrong, sympy just starts returning wrong results even in cases
which were ok before.

There are two ways to fix it. Improve the _get_cache_index() method, or rewrite
the Order.contains() not to use cache. I chose the second method. It turns out
that the fix is really easy and nice once you know what to do.

Timings using "py.test sympy/series/"

Before: 22.51
After: 23.04

The timings vary +- 0.5s, but the patch seems to slow things down by 0.5s in
average. That is imho acceptable, considering that the code is now more clean
and robust.

Imho there should be no caching in Order, just maybe in limits, but even there
I think the easy cases should be made fast by not calling gruntz, but wiring in
the results. We are doing it this way already and I think we should just
continue this path.

So I continued and in the next patches I removed caching from Order completely
and it not only made the code a lot simpler, but series tests are now 1.5x
faster!!

The lesson learned: doing caching in hackish way is evil. Don't do that.