[8/24]: rework assumptions to use FactRules  (no .assume())
.assume() is now disallowed -- in fact it was never a good idea to change
assumptions in-situ.

When we'll refactor them to be separate of expressions, it could make sense,
but even then a lot of assumptions would be still bound to objects (e.g.
Rational.is_integer=False for performance reasons)

So let's just remove one tiny place where .assume() was 'used' to modify
assumptions in-situ -- the only test.


Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Mateusz Paprocki <mattpap@gmail.com>