[4/24]: rework assumptions.py to use FactRules   (kill ._eval_is_nonnegative)
As we now use inference engine with proper beta-rules support (&(a,b) -> c),
there is no need to provide ad-hock things like generic ._eval_is_nonnegative()

In fact this old ._eval_is_nonnegative() was wrong -- it was correct only when
an object in hand was real.

So let's get rid of old unneeded stuff and use generic inference.


Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Mateusz Paprocki <mattpap@gmail.com>