[5/24]: rework assumptions to use FactRules  (BasicMeta)
Rework BasicMeta to deduce all possible consequence of a class
default_assumptions and make it complete at class creation time.

e.g. if we define:

    class EvenInteger(Basic):
        is_integer  = True
        is_even     = True

        ...

it will infer:

    integer -> real -> complex
    even    -> !odd
    (and probably other facts)

and store them as class.default_assumptions.

----

Why is this needed?

In previous assumptions, when an object was created, Basic.__new__ was calling
.assume() and the same work was done again and again.

It's better to precompute whatever is possible, and we do exactly this -- so
now cls.default_assumptions stores convex hull of all cls initial premises.


Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Mateusz Paprocki <mattpap@gmail.com>