risch: "robustify" test_heurisch_fractions()
When integrating depending on tests invokation order it either ruturns

    5*log(+2*x**6 - 5) / 12,     or
    5*log(-2*x**6 + 5) / 12


Since

    log(-x) = i + log(x)

both integration results are still correct, but I'd suspect something is wrong here.

Mateusz, I need your help to track this -- this does not depend on caching be
turned on or off.

Mateusz:

"""
Yes, both results are correct but we must finally do something to fix
this annoying behavior (I mean playing with minus sign in expressions
and dependence on term order which, in case of expressions, is almost
always irrevelant).

+1

Thanks,
Mateusz
"""


Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Mateusz Paprocki <mattpap@gmail.com>