Improve output and tests of ReprPrinter + some basic fixes for printing.
The following is included:
       (1) Excessive tests for ReprPrinter
       (2) Improvements for ReprPrinter
       (3) Fix of PythonPrinter (necessary because of slightly changed output
of ReprPrinter)
       (4) Add _print_Zero to StrPrinter

Most important, in my opinion, is that all tests of ReprPrinter now also
test if the Relation eval(srepr(expr))==expr holds. E.g. this revealed
that (at least for the tested examples) the srepr of Reals must not
include more digits than their precision.

Signed-off-by: Sebastian Krämer <basti.kr@gmail.com>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>