pretty: merge handling of str and unicode
Since str and unicode are both subclasses of basestring we can merge them on
this basis and handle just basestring.

Here is why:

Let's recall how printing works (see Printer docs) -- it sequentially tries
_print_<class>, _print_<class_base>, _print_<class_base_base> etc...

so here is what Printer tries for str:

    _print_str
    _print_basestring

and for unicode:

    _print_unicode
    _print_basestring

And since we have that _print_basestring now, everything works just fine, but
with less code.


Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>