Move Add.compare_terms to Basic._compare_pretty
This is currently used for pretty-sorting when printing Adds, but since we are
going to do the same for dicts and sets, it is convenient to move this
pretty-comparison to Basic.

Leading underscore is added to the name to indicate that this function is only
capable to deal with Basic objects. To compare both SymPy and non-SymPy objects
we'll add Basic.compare_pretty in the next patch.


Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>