printing: `srepr` sisters -- `sstr` and `sstrrepr`
Like we have `srepr` to obtain the result of printing an expression through
ReprPrinter, it would be handy to have `sstr` do the same with StrPrinter.

Also, introduced is a new function `sstrrepr` which job is to print everything
strings, except strings itself which are printed in quotes, e.g.

>>> e = ['a', 'b', 'c', x]
>>> print sstr(e)
[a, b, c, x]

>>> print sstrrepr(e)
['a', 'b', 'c', x]


That's why `sstrrepr` could be useful for hooking into sys.displayhook


Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>