tsolve: teach it to solve  exp(x)+exp(-x)==y  (#1011)
If we consider the following equation

    f(x)+g(x)+... = y

This is done by simply trying to substitute lhs f(x) -> t, and see what
happens.


The interesting case here is when all the functions in the lhs part go away --
we can now proceed with solving algebraic equation and then do back
substiotution.


Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>