Make 1+O(1) return O(1) instead of 1+O(1).
The problem was here:

In [4]: Add.flatten((sympify(1), O(1)))
Out[4]: ([1, O(1)], [], None)

In [5]: Add.flatten((sympify(x), O(1)))
Out[5]: ([O(1)], [], None)

[5] is right, but [4] is wrong, it should return the same as [5].
The reason for that is that O(1).contains(1) was not called (and action taken
depending on the result) in Add.flatten(). This was a mistake caused by the
fact that numbers are handled separately.

This is fixed by this patch and a tests written.