latex(1/x) should return $\frac{1}{x}$  (#1030)
when I try something like :

>>> latex(1/x)

the answer is :

'${x}^{(-1)}$'

and not $\frac{1}{x}$.

Is it a way to get the previous result ?

Ondrej:

Thanks a lot for the patch.

I am little afraid of its fragility, i.e. what happens if someone changes the way
-1*x is printed, e.g. from:

In [3]: latex(-1/x)
Out[3]: '$- \\frac{1}{x}$'

to

In [3]: latex(-1/x)
Out[3]: '$-\\frac{1}{x}$'

and your code will stop working.

How about just returning \\frac{1}{x} directly in your code?

...

If there is not a better way to do it, I am ok with it.

Signed-off-by: Sebastian Kreft <skreft@gmail.com>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>