Fix Integer(Integer(4))  (#1034)
When we do

    sympify('Integer(4)')

it boils down to

    Integer(Integer(4))

because internally our sympify converts Python ints to Rationals at
Python AST level.

But previously Integer ctor was acception only Python ints and longs, so
let's fix it.

Also, when doing so we should always remember that Integer(n) should be
optimized for cache-hit case. Why? -- The answer is in the second patch.

Signed-off-by: Kirill Smelkov <kirr@landau.phys.spbu.ru>
Signed-off-by: Ondrej Certik <ondrej@certik.cz>